dsmeLayer/TimerMultiplexer: improve error check for timer in the past
All timer related logic is meant to work with a counter that wraps around. The timer multiplexer however asserted on a target time that is smaller than the current time, which would happen when setting a timir that spans across the wrap. This commit changes the check based on the assumption that the timer will never be used for long (>half period) timeouts and treats this as an invalid value. This also covers an actual overflow, where the difference would then be very large